perf: use ISSG for Puck pages
* feat: purge next cache on POST

* perf: purge Next.js supabase Data cache

Next.js globally patches fetch, which the supabase-js package uses under the hood. This change patches the fetch implementation used by supabase and adds a cache tag based on the path, which we purge on publish. Alternatively we could use the REST API to connect to Supabase, but I had a hard time getting this working.

* build: return null if data is undefined

---------

Co-authored-by: Chris Villa <chrisvilla@me.com>